requirements.txt
python-dotenv
fastapi
uvicorn
ultralytics
opencv-python-headless
python-multipart
pillow
numpy
